requirements.txt
Django==5.0.3
djangorestframework==3.15.1
mysqlclient==2.2.4  
python-decouple==3.8 
